PatternFolds

PatternFolds.MVectorFold

PatternFolds.PatternFold

PatternFolds.VectorFold

Base.eltype

Base.iterate

Base.length

Base.rand

PatternFolds.folds

PatternFolds.gap

PatternFolds.pattern

PatternFolds.pattern

PatternFolds.pattern_length

PatternFolds.reset_pattern!

PatternFolds.set_fold!

PatternFolds.unfold

PatternFolds.MVectorFold — Type

MVectorFold{T,V <: AbstractVector{T}}

A mutable structure for folded vector that extends the methods of AbstractVector. Compared to VectorFold, this tructure is about 20% faster using iterators. Unfolding is twice slower though. Note that this structure keep an active pointer to the current unfolded pattern. However, its external behavior is similar to VectorFold.

PatternFolds.PatternFold — Type

PatternFold{T, P}

An abstract stype used as an interface for folded patterns such as VectorFold. To implement the interface and inherit from it, a new structure must define three fields:

pattern::P. Note that both length(::P) and rand(::P) methods must be available

gap::T

folds::int

PatternFolds.VectorFold — Type

VectorFold{T,V <: AbstractVector{T}}

A folded vector structure that extends the methods of AbstractVector to a folded structure.

Base.eltype — Method

eltype(pf<: PatternFolds)

Base.iterate — Method

iterate(iter)

Extends iterate methods from Base to allow forward and reverse iteration on both VectorFold and MVectorFold.

Base.length — Method

length(pf<:PatternFold)

Return the length of pf if unfolded.

Base.rand — Method

rand(pf<:PatternFold)

Returns a random value of pf as if it was unfolded.

PatternFolds.folds — Method

folds(<:PatternFold)

Return the number of folds. An infinite folded pattern returns 0.

PatternFolds.gap — Method

gap(<:PatternFold)

Return the gap between the starts of consecutive folds.

PatternFolds.pattern — Method

pattern(vf, index)

Return the element at index in the original pattern.

PatternFolds.pattern — Method

pattern(<:PatternFold)

Return the pattern of any PatternFold. The pattern defines the occurences of the first fold.

PatternFolds.pattern_length — Method

pattern_length(pf<:PatternFold)

Return the length of the basic pattern of pf.

PatternFolds.reset_pattern! — Method

reset_pattern!(<:PatternFold)

Reset the unfolded pattern to the first fold.

PatternFolds.set_fold! — Function

set_fold!(mvf::MVectorFold, new_fold = mvf.current + 1)

Set the unfolded pattern to new_fold. By default move the next fold after current.

PatternFolds.unfold — Method

unfold(vf::VectorFold; from=1, to=folds(vf))

Construct the unfolded version of vf (with the same type as pattern(vf)) based. Please note that using an iterator on vf avoid memory allocation, which is not the case of unfold.